refactor: further separate out inference and hooks
